sort indexer-serverless exports
